spi: spi-fsl-dspi: Choose pinctrl base on pinctrl-names
commit 718a912f7fcd8c841d9bd574b0bfa445db16582d from
https://source.codeaurora.org/external/autobsps32/linux

The spi controller node in dts cand define both master
and slave pinctrl. The selection between them will be done based
on pinctrl-names. The default pinctrl will be loaded first
and will be used by the master. If the controller is configured
as slave (spi-slave property is added in the dts node), the
driver will look for the "slave" pinctrl and apply it if
found.

Issue:  ALB-1129
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>