linux-s32: pcie: Fix suspend/resume crash with no EP plugged
commit 098454a8fbe16a5b3297c169c30c4a24d2447a94 from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

Since we always probe the PCIe controller during Linux boot,
we should allow the suspend/resume routines to be executed with
no EP pluged, hence no linkup.
Before this patch, in 'suspend' callback we checked if we had
link-up, and in a negative case we exited the callback which lead
to a Linux crash.

Issue: ALB-7869
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>