pcie:s32:dma:device-tree: Fix DMA interrupt on R45 for PCIe1
commit bd495f4308cd95fe5c78b680b4274f53f318dc71 from
https://source.codeaurora.org/external/autobsps32/linux

The DMA interrupt was not triggered on R45 for PCIe1 because the IDs
for S32G were used in the device tree for R45 as well.
For SerDes1, the interrupt IDs are different between S32G and R45.
This fix splits the interrupt definitions and fixes the IDs for R45.

Issue: ALB-4548

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>