pinctrl: s32cc: Place each group into a function block
commit 77429ccdcbd9f130c9b0562433554659fd9d7872 from
https://source.codeaurora.org/external/autobsps32/linux

All S32CC pinmuxing nodes are defines as:
    &pinctrl {
      <soc> {
        can0 {
          fsl,pins = <...>;
        }
      };

      <board> {
        can0 {
          fsl,pins = <...>;
        }
      };
    };
The <soc> and <board> nodes are interpreted by the driver as
pinmuxing functions, which is wrong because this is at least
inaccurate. This commit places each group in the example above
in a function node. Therefore, each function node will have only
one group attached to it. It may seen unnecessary, but it
helps to transition to generic pinconf interface.

Issue: ALB-8466
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>