s32gen1: qspi: make the last LUT config sharable
There are 80 LUTs in qspi controller, and every 5 LUTs form a sequence
to generate a valid flash transaction. A sequence corresponds to a LUT
config. And the maximum of LUT configs are 16.

Now the implimentation of s32g qspi driver is that one type of cmd
opcode occupies one LUT config. This means that this driver only
operates 16 types of cmd.

We can make the last LUT config which is the 15th to be sharable, the
previous 15 cmd can occupy 15 LUT configs, all the left cmds share the
15th LUT config.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>