pinctrl: s32: rename "pins" property
commit fbdb16a86eaa1d2d1f6796748c9fb7f970cf819e from
https://source.codeaurora.org/external/autobsps32/linux

The "pins" property used inside the pinctrl node was specific to the
driver. Other drivers don't use this property. It specifies the range
of MSCR/IMCR registers that are valid and correspond to the "reg"
property.

Change it into "nxp,pins". Also, remove the phandle reference from it
since it is always referencing the parent pinctrl node.

Issue: ALB-9022

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>