pinctrl: s32: split memory region for S32G SIUL21 MSCRs
commit 8985e08b3e7b3bdb579284b01f348104b6287bd7 from
https://source.codeaurora.org/external/autobsps32/linux

There is a gap for the SIUL21 module for S32G SoCs. GPIOs
123-143 are not available. Currently, trying to export a GPIO in that
region would fail because the corresponding `pinctrl_pin_desc` cannot
be found. However, the memory region is still reserved.

The problem would be seen when trying to use the same DTB in U-Boot
where the pin descriptions are not hardcoded.

Split the current memory region for S32G's SIUL21 MSCRs. Update the
available `pins` range for the pinctrl driver.

Issue: ALB-9022

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>