dts: s32g: disable mdio nodes of Aquantia phy for rdb2/3
For RDB2 and RDB3 boards, the Aquantia phy are connected to EMAC of pfe1
interface which is disabled by default, so the mdio nodes of Aquantia phy
also need to be disabled, or else, there will resume error like the following:

 PM: dpm_run_callback(): mdio_bus_phy_resume+0x0/0x110 returns -19
 Aquantia AQR107 stmmac-0:03: PM: failed to resume: error -19

This patch is to fix it. And the node will be "okey" when the pfe1 is enabled.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>