RISC-V: set memblock allocation direction to bottom-up
commit c6e097a133f89dfa8a6e7eeefadee734b00d8ea1 from
https://github.com/linux4microchip/linux.git

RISC-V currently allocates memblocks in a top-down manner.
Unfortunately, the first allocations occur in paging_init(), *before*
we have scanned for reserved memory.

Accordingly, if the topmost memblock is destined to be reserved memory,
we may erroneously allocate from it. At best, this means a memory
reservation will later fail, and at worst will crash the kernel (or
clobber some other application).

To avoid this, begin to allocate from the lowest memblock up.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>