RISC-V: configs: convert generic-service to a module for mpfs
commit 5ef15c87346fdf500b0d73992e00585fe1ba24fc from
https://github.com/linux4microchip/linux.git

Ivan complained about the warning emitted by the generic service driver
as being non-ideal for our default image. To appease him, and general
optics, make the driver a module that must be specifically loaded to
use the system service examples, where the meaning of the warning can
be explained more wordily.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>