RISC-V: configs: enable IO_STRICT_DEVMEM for mpfs
commit 8264bf739fc2af59e161a81412c7f4953b640e5d from
https://github.com/linux4microchip/linux.git

If we're going to be pointing people at the /dev/mem tools for use with
some of the apps demos, we should ensure that (by default at least)
noone is using /dev/mem to fiddle with peripherals that already have
drivers bound.
Enable IO_STRICT_DEVMEM to block those types of access.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>