drm/bridge: microchip-lvds: Replace clk_prepare with prepare_enable
commit 6601c66ee8363b99240f24fc017722e383b3a8c0 from
https://github.com/linux4microchip/linux.git

Remove clk_prepare function in probe and replace the clk_enable() function
in lvds_enable with clk_prepare_enable().

When the lvds controller is probed first the clk_prepare() function
takes ownership of the clock causing issues for other drivers when the
probe, moving the prepare to when the lvds clock is ready to be enabled
fixes this issue.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>