iio: adc: at91-sama5d2_adc: adapt the driver for sama7d65
commit dcfd35014dd1facb07fd749cf80ed828d0e7c21c from
https://github.com/linux4microchip/linux.git

Add support to sama7d65 ADC. The differences are highlighted with the
compatible. The number of channels and the parsing of temperature
calibration data are the main differences. This patch also adds support
to read calibration data from SRAM0 using the phandle. The temperature
sensor init sequence differences are handled by using a different
callback function. The same is added to the platform data of the
corresponding device.

Signed-off-by: Varshini Rajendran <varshini.rajendran@microchip.com>
Reviewed-by: Nicolas Ferre <nicolas.ferre@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>