media: i2c: imx334: revert link frequency configuration
commit 1bed65d4c4b40748004dde59c45cc62e5cd4c9ae from
https://github.com/linux4microchip/linux.git

The imx334 register values reverted for required link frequency, as
per the MIPI CSI IP Configuration in the fabric video pipeline.

The current value is what got upstreamed, but does not work for the link
speed in either the dt-binding example or what is in the existing
overlay for this device.

The correct fix would be to parse the link-frequency from the
endpoint, but an immediate fix is required for the release.

Signed-off-by: Shravan Chippa <shravan.chippa@microchip.com>
[conor: added some more info to the commit message]
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>