Revert "Merge linux-6.6-trunk/riscv/core into linux-6.6-mchp+fpga"
commit f19a79641699ea4400db4704694cf866bdbba9ba from
https://github.com/linux4microchip/linux.git

This branch has a single commit in it, so rather than revert the
commit in question, just revert the merge of the entire branch.

The commit made the riscv kernel allocate memblocks bottom-up rather
than top down, which was done because boot on the discokit (or other
boards with 1 GiB of memory) we had issues with the devicetree being
corrupted. However, it caused boards with NVMe drives connected to PCIe
to fail to boot.

Revert commit 35da41dfe4efb367d35d1551f88bb252e26df850, reversing
changes made to a50ab3c5953920e8ab05ff65b7758284a84db826. Obviously this
will reintroduce the problem with boot on the Discovery kit, but we've
found a different resolution there - at least one that should suffice
while the issue is investigated upstream.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>