mailbox: mpfs: support new syscon based devicetree configuration
commit 8e1c0b8dd575aad583eca73f59157d67b7a31f9e from
https://github.com/linux4microchip/linux.git

The two previous bindings for this hardware were incorrect, as the
control/status and interrupt register regions should have been described
as syscons and dealt with via regmap in the driver. Add support for
accessing these registers using that method now, so that the hwmon
driver can be supported without using auxdev or hacks with io_remap().

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>