irqchip/mpfs-irq-mux: Permit probing as a child of a syscon using a regmap
commit 755357b3aec01ffc4c5ab201805eca6d9d61dd24 from
https://github.com/linux4microchip/linux.git

Support the new method of describing the interrupt mux as a child of a
syscon, and use regmap to access the mux configuration register. This is
a more suitable description as the mux only has a single register in the
mss-top-sysreg syscon.

While mainline will never know about anything other than supporting this
device as a child of a syscon, we need to maintain support in our vendor
kernel for the old method of probing using of_iomap() as a fallback if
the syscon node does not exist.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>