net: dsa: microchip: add suspend/resume support to KSZ DSA driver
commit 4b361a29cc5b093c3bb0589cf181294d06c1c42e from
https://github.com/linux4microchip/linux.git

The KSZ DSA driver starts a timer to read MIB counters periodically to
avoid count overrun.  When the system is suspended the SPI register
write returns an error because the SPI system now displays an error
message when it is suspended.  This patch stops the timer when the driver
is suspended and restarts the timer when resumed.

Signed-off-by: Tristram Ha <tristram.ha@microchip.com>
Tested-by: Hari Prasath Gujulan Elango <Hari.PrasathGE@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>