drm/bridge: fix LVDS controller bus format
commit c561d1072b519ad6053449e1a602e2d46489b6b3 from
https://github.com/linux4microchip/linux.git

The current LVDS controller driver is hardcoded to map LVDS lanes to the
JEIDA format. Consequently, connecting an LVDS display that supports the
VESA format results in a distorted display due to the format mismatch.This
patch resolves the issue by querying the panel driver and setting the
appropriate format accordingly.

Signed-off-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>
Signed-off-by: Charan Pedumuru <charan.pedumuru@microchip.com>
Tested-by: Manikandan Muralidharan <manikandan.m@microchip.com>
Tested-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>
Reviewed-by: Dharma Balasubiramani <dharma.b@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>